refactor: split session icon views
