+++ Sources/SessionIndexIconImage.swift
@@ -1,6 +1,3 @@
-import AppKit
-import CmuxAppKitSupportUI
-import CmuxFoundation
 import SwiftUI
 
 /// Shared AppKit-backed icon view for Vault section headers and previews.
@@ -30,69 +27,5 @@ struct SessionIndexSectionIconImage: View, Equatable {
                 fallbackSource: .workspaceIcon(.folder)
             )
         }
-    }
-}
-
-private struct SessionIndexAgentIconImage: View, Equatable {
-    let agent: SessionAgent
-    let size: CGFloat
-
-    static func == (lhs: Self, rhs: Self) -> Bool {
-        lhs.agent == rhs.agent && lhs.size == rhs.size
-    }
-
-    var body: some View {
-        if let assetName = agent.assetName {
-            CmuxResolvedIconImage(request: CmuxResolvedIconRequest(
-                source: .asset(name: assetName, bundle: .main),
-                size: NSSize(width: size, height: size),
-                fallbackSource: .systemSymbol(
-                    name: "person.crop.circle.fill",
-                    accessibilityDescription: nil
-                ),
-                fallbackTintColor: .secondaryLabelColor
-            ))
-            .frame(width: size, height: size)
-        } else {
-            SessionIndexResolvedSystemSymbolImage(
-                systemName: agent.systemImageName ?? "person.crop.circle",
-                pointSize: max(size - 2, 10),
-                size: size,
-                weight: .regular,
-                tintColor: .secondaryLabelColor,
-                fallbackSource: .systemSymbol(
-                    name: "person.crop.circle.fill",
-                    accessibilityDescription: nil
-                )
-            )
-        }
-    }
-}
-
-private struct SessionIndexResolvedSystemSymbolImage: View {
-    @Environment(\.cmuxGlobalFontMagnificationPercent) private var globalFontPercent
-
-    let systemName: String
-    let pointSize: CGFloat
-    let size: CGFloat
-    let weight: NSFont.Weight
-    let tintColor: NSColor
-    let fallbackSource: CmuxResolvedIconSource?
-
-    var body: some View {
-        let rasterSize = GlobalFontMagnification.scaledSize(pointSize, percent: globalFontPercent)
-        CmuxResolvedIconImage(request: CmuxResolvedIconRequest(
-            source: .systemSymbol(name: systemName, accessibilityDescription: nil),
-            size: NSSize(width: rasterSize, height: rasterSize),
-            tintColor: tintColor,
-            symbolWeight: weight,
-            fallbackSource: fallbackSource
-        ))
-        // Keep the magnified raster's own layout size, then center it in the
-        // design-size slot. This mirrors `CmuxSystemSymbolImage(magnified:)`
-        // and prevents the AppKit image view from scaling the larger bitmap
-        // back down to the unscaled slot.
-        .frame(width: rasterSize, height: rasterSize)
-        .frame(width: size, height: size)
     }
 }
